Check cache before querying the api for searches
Moved all the search callbacks out of the Search component back to the GodDirectory. The Search box knew a little too much regarding the cache fallback etc. Felt it was more the responsibility of the GodDirectory to determine how we handle this. The Search component now just passes its value back to a callback to do whatever it wants with it.